Move all files on pre_setup.py to libs
Simplify code and prepare for prep release 0.2
