+++ src/main/python/systemds/utils.py
@@ -38,6 +38,7 @@ def set_auto_start_jmlc(should_start_server: bool) -> None:
     """
     Can activate/deactivate automatic start of JMLC API Gateway.
     Has to be called before any communication took place (`.compute()` was called).
+
     :param should_start_server: activate/deactivate
     """
     global AUTO_START_JMLC_SERVER
@@ -48,24 +49,17 @@ def get_gateway() -> JavaGateway:
     """
     Gives the gateway with which we can communicate with the SystemDS instance running a
     JMLC (Java Machine Learning Compactor) API.
+
     :return: the java gateway object
     """
     global JAVA_GATEWAY
     if JAVA_GATEWAY is None:
         if AUTO_START_JMLC_SERVER:
-            separator = ':'
-            # note that windows has a different separator for classpath
-            if platform.system() == 'Windows':
-                separator = ';'
-            files = _get_jar_file_paths()
-            if any((fnmatch(file, JAR_FILE_STANDALONE) for file in files)):
-                classpath = separator.join(files)
-            else:
-                libs = os.path.join(_get_module_dir(), 'systemds-java', 'lib', '*')
-                classpath = separator.join([libs] + files)
-            process = subprocess.Popen(
-                ['java', '-cp', classpath, 'org.tugraz.sysds.pythonapi.PythonDMLScript'], stdout=subprocess.PIPE)
+            classpath = os.path.join(_get_module_dir(), 'systemds-java', 'lib', '*')
+            process = subprocess.Popen(['java', '-cp', classpath, 'org.tugraz.sysds.pythonapi.PythonDMLScript'],
+                                       stdout=subprocess.PIPE)
             process.stdout.readline()  # wait for 'Gateway Server Started\n' written by server
+            assert process.poll() is None, "Could not start JMLC server"
         JAVA_GATEWAY = JavaGateway()
     return JAVA_GATEWAY
 
@@ -74,6 +68,7 @@ def create_params_string(unnamed_parameters: Iterable[str], named_parameters: Di
     """
     Creates a string for providing parameters in dml. Basically converts both named and unnamed parameter
     to a format which can be used in a dml function call.
+
     :param unnamed_parameters: the unnamed parameter variables
     :param named_parameters: a dictionary of parameter names and variable names
     :return: the string to represent all parameters
@@ -85,20 +80,8 @@ def create_params_string(unnamed_parameters: Iterable[str], named_parameters: Di
 def _get_module_dir() -> os.PathLike:
     """
     Gives the path to our module
+
     :return: path to our module
     """
     spec = find_spec(MODULE_NAME)
     return spec.submodule_search_locations[0]
-
-
-def _get_jar_file_paths() -> List[str]:
-    """
-    Returns paths to SystemDS jar files
-    :return: paths to SystemDS jar files
-    """
-    java_dir = os.path.join(_get_module_dir(), 'systemds-java')
-    jar_file_names = []
-    for file in os.listdir(java_dir):
-        if any((fnmatch(file, name) for name in JAR_FILE_NAMES)):
-            jar_file_names.append(os.path.join(java_dir, file))
-    return jar_file_names
